Fixes after running performance tests

+++ app/repositories/projects_repository.rb
@@ -2,17 +2,17 @@ class ProjectsRepository < BaseRepository
   def create(team_id, attributes)
     project = asana_client.create_project(
       A9n.asana[:workspace_id], team_id, attributes)
-    collection.add(project)
+    collection.add(project) if project
   end
 
   def update(project_id, attributes)
     project = asana_client.update_project(project_id, attributes)
-    project.update(attributes)
+    project.update(attributes) if project
   end
 
   def delete(project_id)
-    asana_client.delete_project(project_id)
-    collection.delete(project_id)
+    success = asana_client.delete_project(project_id)
+    collection.delete(project_id) if success
   end
 
   private
